Replaced mastercomfig 7 alpha with mc 6.14.1
Lots of little changes again, can't remember most of em.
This is the windows config.
mastercomfig 6.14.1 has some different requirements for custom
configs and so a lot of changes is just complying to those changes
in cfg/user/disabled is a lot of legacy scripts for miscellaneous
things. At the moment the cfg/user script is all of my main config
settings all packaged up neatly in that folder.